chore: refactor node style
