+++ packages/mind/src/utils/index.ts
@@ -3,7 +3,7 @@ export * from './weak-maps';
 export * from './mind';
 export * from './layout';
 export * from './shape';
-export * from './colors';
+export * from './node-style';
 export * from './is-virtual-key';
 export * from './draw-placeholder';
 export * from './direction-detector';
